pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nikame"
version = "1.3.1"
description = "NIKAME: High-fidelity AI & Cloud Infrastructure Scaffolding Framework (Matrix Engine)"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "omdeepb69" },
]
keywords = [
    "mlops", "infrastructure", "ai", "rag", "scaffolding", "automation", 
    "devops", "llm", "matrix-engine", "fastapi", "docker", "kubernetes"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1",
    "questionary>=2.0",
    "pydantic[email]>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "networkx>=3.1",
    "rich>=13.0",
    "docker>=7.0",
    "httpx>=0.25",
    "anyio>=4.0",
    "tenacity>=8.0",
    "structlog>=24.0",
    "tomli-w>=1.0.0",
    "thefuzz[speedup]>=0.20.0",
    "watchfiles>=0.21.0",
    "pyngrok>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/nikame-dev/nikame"
Repository = "https://github.com/nikame-dev/nikame"
Issues = "https://github.com/nikame-dev/nikame/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "types-pyyaml",
    "types-docker",
]
cloud = [
    "boto3>=1.29",
    "kubernetes>=28.0",
]
ml = [
    "huggingface_hub>=0.19",
    "litellm>=1.10",
    "streamlit>=1.30",
    "requests>=2.31",
]
all = [
    "boto3>=1.29",
    "kubernetes>=28.0",
    "huggingface_hub>=0.19",
    "litellm>=1.10",
    "streamlit>=1.30",
    "requests>=2.31",
]

[project.scripts]
nikame = "nikame.cli.main:cli"

[project.entry-points."nikame.plugins"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "DTZ", "T20", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401", "S101", "E501", "S603", "S607", "S701", "B904", "RUF012", "S104", "B007", "S105", "F401", "E402", "W291", "B006", "E701", "F841", "W293", "RUF013"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build]
include = [
  "nikame/**/*.py",
  "nikame/**/*.j2",
  "nikame/**/*.yaml",
  "nikame/**/*.json",
  "nikame/**/*.md",
  "nikame/**/*.sh",
]

[tool.hatch.build.targets.wheel]
packages = ["nikame"]
